refactor: move _make_shape_path and path builders from shape to _shape
