+++ labelme/_shape.py
@@ -1,12 +1,14 @@
 from __future__ import annotations
 
 from collections.abc import Iterable
+from typing import Final
+from typing import Protocol
 
 from PyQt5 import QtCore
+from PyQt5 import QtGui
 
 from . import utils
 from .shape import Shape
-from .shape import _make_shape_path
 
 
 def _argmin(values: Iterable[float]) -> tuple[int, float] | None:
@@ -71,3 +73,46 @@ def contains_point(*, shape: Shape, point: QtCore.QPointF) -> bool:
 def bounds(*, shape: Shape) -> QtCore.QRectF:
     path = _make_shape_path(shape_type=shape.shape_type, points=shape.points)
     return path.boundingRect()
+
+
+def _build_rectangle_path(*, points: list[QtCore.QPointF]) -> QtGui.QPainterPath:
+    out = QtGui.QPainterPath()
+    if len(points) == 2:
+        out.addRect(QtCore.QRectF(points[0], points[1]))
+    return out
+
+
+def _build_circle_path(*, points: list[QtCore.QPointF]) -> QtGui.QPainterPath:
+    out = QtGui.QPainterPath()
+    if len(points) == 2:
+        radius = utils.distance(points[0] - points[1])
+        out.addEllipse(points[0], radius, radius)
+    return out
+
+
+def _build_polyline_path(*, points: list[QtCore.QPointF]) -> QtGui.QPainterPath:
+    out = QtGui.QPainterPath()
+    if not points:
+        return out
+    out.moveTo(points[0])
+    for vertex in points[1:]:
+        out.lineTo(vertex)
+    return out
+
+
+class _PathBuilder(Protocol):
+    def __call__(self, *, points: list[QtCore.QPointF]) -> QtGui.QPainterPath: ...
+
+
+_PATH_BUILDERS: Final[dict[str, _PathBuilder]] = {
+    "rectangle": _build_rectangle_path,
+    "mask": _build_rectangle_path,
+    "circle": _build_circle_path,
+}
+
+
+def _make_shape_path(
+    *, shape_type: str, points: list[QtCore.QPointF]
+) -> QtGui.QPainterPath:
+    builder = _PATH_BUILDERS.get(shape_type, _build_polyline_path)
+    return builder(points=points)
